Better category handling
Add arguments that specify how many categories and tags to associate
with posts (vs. just capping it at 5).

Use the iterator index when adding to $this->tags and $this->categories
rather than a default 0-based index, as otherwise we get mismatched
slugs/names between the listing of terms at the top of the export file
and what's actually associated with the posts.

Also a minor whitespace change while I'm at it.